+++ tests/fuzzylevenshtein/test_fuzzysbs.py
@@ -1,35 +1,4 @@
-import pandas as pd
-import pytest
-
 from wookie.comparators import FuzzyWuzzySbsComparator
-from wookie.pandasconnectors import CartesianConnector
-
-
-@pytest.fixture
-def df_circus():
-    left = pd.DataFrame(
-        {
-            'name': [
-                'hello world',
-                'hello big world',
-                'holy grail',
-                'holy moly'
-            ]
-        }
-    )
-    left.index.name = 'ix'
-    right = left.copy()
-    right['ix'] = pd.Series(['a', 'b', 'c', 'd'], index=right.index)
-    right.set_index('ix', drop=True, inplace=True)
-    X = [left, right]
-    return X
-
-
-@pytest.fixture
-def circus_sbs(df_circus):
-    cc = CartesianConnector(ixname='ix', lsuffix='left', rsuffix='right', on='name')
-    X_sbs = cc.showpairs(X=df_circus)
-    return X_sbs
 
 
 def test_cclr(df_circus, circus_sbs):
