+++ tests/leftrightsbspipe/test_lrsbspipe.py
@@ -1,46 +1,11 @@
 import pandas as pd
-import pytest
 from sklearn.pipeline import make_union, make_pipeline, Pipeline
 
 from wookie.comparators import FuzzyWuzzySbsComparator
 from wookie.pandasconnectors import VectorizerConnector, ExactConnector, CartDataPasser
-from wookie.preutils import concatixnames
+
 
 # TODO: this test is to test the possibility of having Left -Right comparators followed by Left-Right Fuzzy comparator
-
-n_lines = 10
-
-
-@pytest.fixture
-def ix_names():
-    ixname = 'ix'
-    lsuffix = 'left'
-    rsuffix = 'right'
-    ixnameleft, ixnameright, ixnamepairs = concatixnames(
-        ixname=ixname, lsuffix=lsuffix, rsuffix=rsuffix
-    )
-    names = dict()
-    names['ixname'] = ixname
-    names['ixnameleft'] = ixnameleft
-    names['ixnameright'] = ixnameright
-    names['ixnamepairs'] = ixnamepairs
-    names['lsuffix'] = lsuffix
-    names['rsuffix'] = rsuffix
-    return names
-
-
-@pytest.fixture
-def df_left():
-    left = pd.read_csv('/Users/paulogier/81-GithubPackages/wookie/operations/data/left.csv', index_col=0,
-                       dtype=str).sample(n_lines)
-    return left
-
-
-@pytest.fixture
-def df_right():
-    right = pd.read_csv('/Users/paulogier/81-GithubPackages/wookie/operations/data/right.csv', index_col=0,
-                        dtype=str).sample(n_lines)
-    return right
 
 
 def test_loaddata(ix_names, df_left, df_right):
